add `rest` example
In the step immediately following this in the outline, we refer to `rest` as something that was previously shown, however its not actually referred to before that point. This minor edit just adds a `rest` example. 

I'm unsure if this was an oversight or a deliberate choice, but it stuck out to me.