Added warning to Light Table install procedure
If the Light Table zip file is unzipped to a location with spaces in the file path, then commands like "Instarepl: open a clojure instarepl." display an error similar to "Error: Unable to access jarfile "C:\Folder With Spaces\...\lein-light-standalone.jar". Removing spaces from the path (or moving the Light Table files) resolves the issue.